test(grid): move kayBoard test in a separate file #2975
