+++ projects/igniteui-angular/src/lib/grids/grid/grid.groupby.spec.ts
@@ -44,41 +44,6 @@ describe('IgxGrid - GroupBy', () => {
         }).compileComponents();
     }));
 
-    function expandCollapceGroupRow(fix: ComponentFixture<DefaultGridComponent>,
-                                    groupRow: IgxGridGroupByRowComponent,
-                                    cell: IgxGridCellComponent) {
-        return new Promise(async (resolve, reject) => {
-            expect(groupRow.focused).toBe(true);
-            expect(groupRow.nativeElement.classList.contains('igx-grid__group-row--active')).toBe(true);
-            if (cell != null) {
-                expect(cell.selected).toBe(true);
-            }
-
-            groupRow.nativeElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'arrowleft', code: 'arrowleft', altKey: true }));
-            await wait(300);
-            fix.detectChanges();
-
-            expect(groupRow.expanded).toBe(false);
-            expect(groupRow.focused).toBe(true);
-            expect(groupRow.nativeElement.classList.contains('igx-grid__group-row--active')).toBe(true);
-            if (cell != null) {
-                expect(cell.selected).toBe(true);
-            }
-
-            groupRow.nativeElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'arrowright', code: 'arrowright', altKey: true }));
-            await wait(100);
-            fix.detectChanges();
-
-            expect(groupRow.expanded).toBe(true);
-            expect(groupRow.focused).toBe(true);
-            expect(groupRow.nativeElement.classList.contains('igx-grid__group-row--active')).toBe(true);
-            if (cell != null) {
-                expect(cell.selected).toBe(true);
-            }
-            resolve();
-        });
-    }
-
     function checkGroups(groupRows, expectedGroupOrder, grExpr?) {
         // verify group rows are sorted correctly, their indexes in the grid are correct and their group records match the group value.
         let count = 0;
@@ -185,7 +150,7 @@ describe('IgxGrid - GroupBy', () => {
         tick();
         fix.detectChanges();
         grid.groupBy({ fieldName: 'ReleaseDate', dir: SortingDirection.Asc, ignoreCase: false,
-            strategy: DefaultSortingStrategy.instance() });
+        strategy: DefaultSortingStrategy.instance() });
         fix.detectChanges();
 
         groupRows = grid.groupsRowList.toArray();
@@ -493,7 +458,7 @@ describe('IgxGrid - GroupBy', () => {
         const grid = fix.componentInstance.instance;
         grid.primaryKey = 'ID';
         fix.detectChanges();
-         grid.groupBy([
+        grid.groupBy([
             { fieldName: 'Released', dir: SortingDirection.Desc, ignoreCase: false, strategy: DefaultSortingStrategy.instance() },
             { fieldName: 'ReleaseDate', dir: SortingDirection.Desc, ignoreCase: false, strategy: DefaultSortingStrategy.instance() },
             { fieldName: 'ProductName', dir: SortingDirection.Asc, ignoreCase: false, strategy: DefaultSortingStrategy.instance() },
@@ -619,8 +584,8 @@ describe('IgxGrid - GroupBy', () => {
         fix.componentInstance.enableSorting = true;
         fix.detectChanges();
         grid.groupBy({ fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false,
-            strategy: DefaultSortingStrategy.instance() });
-        fix.detectChanges();
+        strategy: DefaultSortingStrategy.instance() });
+    fix.detectChanges();
 
         let groupRows = grid.groupsRowList.toArray();
         let dataRows = grid.dataRowList.toArray();
@@ -738,484 +703,7 @@ describe('IgxGrid - GroupBy', () => {
         expect(grid.groupingExpressions.length).toEqual(1);
     }));
 
-    // GroupBy + Selection integration
-    it('should toggle expand/collapse state of group row with ArrowRight/ArrowLeft key.', async(() => {
-        const fix = TestBed.createComponent(DefaultGridComponent);
-        const grid = fix.componentInstance.instance;
-        fix.componentInstance.width = '400px';
-        fix.detectChanges();
 
-        grid.groupBy({ fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false,
-            strategy: DefaultSortingStrategy.instance() });
-        fix.detectChanges();
-        const gRow = grid.groupsRowList.toArray()[0];
-        expect(gRow.expanded).toBe(true);
-        const evtArrowLeft = new KeyboardEvent('keydown', { key: 'ArrowLeft', altKey: true });
-
-        const evtArrowRight = new KeyboardEvent('keydown', { key: 'ArrowRight', altKey: true });
-        gRow.element.nativeElement.dispatchEvent(evtArrowLeft);
-        fix.detectChanges();
-
-        expect(gRow.expanded).toBe(false);
-
-        gRow.element.nativeElement.dispatchEvent(evtArrowRight);
-        fix.detectChanges();
-        expect(gRow.expanded).toBe(true);
-    }));
-
-    it(`focus should stays over the group row when expand/collapse
-        with ArrowRight/ArrowLeft keys and grid is scrolled to bottom`, (async () => {
-        const fix = TestBed.createComponent(DefaultGridComponent);
-        const grid = fix.componentInstance.instance;
-        fix.componentInstance.width = '600px';
-        fix.componentInstance.height = '500px';
-        await wait(30);
-        fix.detectChanges();
-
-        grid.groupBy({ fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false,
-            strategy: DefaultSortingStrategy.instance() });
-        fix.detectChanges();
-
-        grid.verticalScrollContainer.scrollTo(grid.verticalScrollContainer.igxForOf.length - 1);
-        await wait(100);
-        fix.detectChanges();
-
-        const groupRows = grid.nativeElement.querySelectorAll('igx-grid-groupby-row');
-        let lastGroupRow = groupRows[groupRows.length - 1];
-        const lastGroupRowIndex = parseInt(lastGroupRow.dataset.rowindex, 10);
-        lastGroupRow.dispatchEvent(new FocusEvent('focus'));
-        await wait(30);
-        fix.detectChanges();
-
-        expect(lastGroupRow.classList.contains('igx-grid__group-row--active')).toBeTruthy();
-        lastGroupRow.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowLeft', altKey: true }));
-
-        await wait(100);
-        fix.detectChanges();
-        lastGroupRow =  grid.nativeElement.querySelector(`igx-grid-groupby-row[data-rowindex="${lastGroupRowIndex}"]`);
-        expect(lastGroupRow).toBeDefined();
-        expect(lastGroupRow.classList.contains('igx-grid__group-row--active')).toBeTruthy();
-        expect(lastGroupRow.getAttribute('aria-expanded')).toBe('false');
-    }));
-
-    it(`should be able to navigate down to the next row when expand the last group row
-    and grid is scrolled to bottom`, (async () => {
-        const fix = TestBed.createComponent(DefaultGridComponent);
-        const grid = fix.componentInstance.instance;
-        fix.componentInstance.width = '600px';
-        fix.componentInstance.height = '500px';
-        await wait(30);
-        fix.detectChanges();
-
-        grid.groupBy({ fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false,
-            strategy: DefaultSortingStrategy.instance() });
-        fix.detectChanges();
-
-        grid.verticalScrollContainer.scrollTo(grid.verticalScrollContainer.igxForOf.length - 1);
-        await wait(100);
-        fix.detectChanges();
-
-        grid.groupsRowList.last.toggle();
-        await wait(30);
-        fix.detectChanges();
-        expect(grid.groupsRowList.last.expanded).toBeFalsy();
-
-        grid.groupsRowList.last.toggle();
-        await wait(30);
-        fix.detectChanges();
-        expect(grid.groupsRowList.last.expanded).toBeTruthy();
-
-        const groupRowIndex = grid.groupsRowList.last.index;
-        UIInteractions.triggerKeyDownEvtUponElem('ArrowDown', grid.groupsRowList.last.nativeElement, true);
-        await wait(100);
-        fix.detectChanges();
-
-        const selectedCell = grid.nativeElement.querySelector('.igx-grid__td--selected');
-        expect(selectedCell).toBeDefined();
-        expect(parseInt(selectedCell.dataset.rowindex, 10)).toBe(groupRowIndex + 1);
-        expect(parseInt(selectedCell.dataset.visibleindex, 10)).toBe(0);
-
-    }));
-
-    xit('should allow keyboard navigation through group rows.', (async () => {
-        const fix = TestBed.createComponent(DefaultGridComponent);
-        const grid = fix.componentInstance.instance;
-
-        fix.componentInstance.width = '400px';
-        fix.componentInstance.height = '300px';
-        grid.columnWidth = '200px';
-        await wait();
-        fix.detectChanges();
-
-        grid.groupBy({ fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false,
-            strategy: DefaultSortingStrategy.instance() });
-        grid.groupBy({ fieldName: 'Released', dir: SortingDirection.Desc, ignoreCase: false, strategy: DefaultSortingStrategy.instance() });
-        fix.detectChanges();
-        await HelperUtils.navigateVerticallyToIndex(grid, 0, 9);
-
-        let row = grid.getRowByIndex(9);
-        expect(row instanceof IgxGridRowComponent).toBe(true);
-        expect(row.focused).toBe(true);
-        expect(row.cells.toArray()[0].selected).toBe(true);
-
-        await HelperUtils.navigateVerticallyToIndex(grid, 9, 0);
-
-        row = grid.getRowByIndex(0);
-        expect(row instanceof IgxGridGroupByRowComponent).toBe(true);
-        expect(row.focused).toBe(true);
-    }));
-
-    xit('should persist last selected cell column index when navigation down through group rows.', async () => {
-        const fix = TestBed.createComponent(DefaultGridComponent);
-        const grid = fix.componentInstance.instance;
-        fix.componentInstance.width = '400px';
-        fix.componentInstance.height = '300px';
-        grid.columnWidth = '200px';
-        fix.detectChanges();
-
-        grid.groupBy({ fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false,
-            strategy: DefaultSortingStrategy.instance() });
-        grid.groupBy({ fieldName: 'Released', dir: SortingDirection.Desc, ignoreCase: false, strategy: DefaultSortingStrategy.instance() });
-        fix.detectChanges();
-
-        grid.parentVirtDir.getHorizontalScroll().scrollLeft = 1000;
-        await wait();
-        let cell = grid.getCellByColumn(2, 'Released');
-        cell.onFocus(new Event('focus'));
-
-        await HelperUtils.navigateVerticallyToIndex(grid, 0, 9, 4);
-
-        grid.markForCheck();
-        fix.detectChanges();
-        const row = grid.getRowByIndex(9);
-        cell = grid.getCellByColumn(9, 'Released');
-        expect(row instanceof IgxGridRowComponent).toBe(true);
-        expect(row.focused).toBe(true);
-        expect(cell.selected).toBe(true);
-    });
-
-    it('keyboard navigation - should focus grouped row when press Tab key and Shift + Tab on a cell', (async () => {
-        const fix = TestBed.createComponent(DefaultGridComponent);
-        const grid = fix.componentInstance.instance;
-
-        fix.componentInstance.width = '600px';
-        fix.componentInstance.height = '600px';
-        grid.columnWidth = '100px';
-        await wait(100);
-        fix.detectChanges();
-
-        grid.groupBy({ fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false,
-            strategy: DefaultSortingStrategy.instance() });
-        fix.detectChanges();
-        let cell = grid.getCellByColumn(2, 'Released');
-        cell.nativeElement.dispatchEvent(new Event('focus'));
-        await wait(100);
-        fix.detectChanges();
-
-        UIInteractions.triggerKeyDownEvtUponElem('tab', cell.nativeElement, true);
-        await wait(100);
-        fix.detectChanges();
-
-        let groupRow = grid.groupsRowList.toArray()[1];
-        cell = grid.getCellByColumn(2, 'Released');
-        await expandCollapceGroupRow(fix, groupRow, cell);
-
-        UIInteractions.triggerKeyDownEvtUponElem('ArrowUp', groupRow.nativeElement, true);
-        await wait(300);
-        fix.detectChanges();
-
-        cell = grid.getCellByColumn(2, 'Released');
-        expect(groupRow.focused).toBe(false);
-        expect(groupRow.nativeElement.classList.contains('igx-grid__group-row--active')).toBe(false);
-        expect(cell.selected).toBe(true);
-        expect(cell.focused).toBe(true);
-
-        cell = grid.getCellByColumn(7, 'Downloads');
-        cell.nativeElement.dispatchEvent(new Event('focus'));
-        fix.detectChanges();
-
-        expect(groupRow.focused).toBe(false);
-        expect(groupRow.nativeElement.classList.contains('igx-grid__group-row--active')).toBe(false);
-        cell.nativeElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'tab', shiftKey: true }));
-        await wait(500);
-        fix.detectChanges();
-
-        groupRow = grid.groupsRowList.toArray()[2];
-        cell = grid.getCellByColumn(7, 'Downloads');
-        await expandCollapceGroupRow(fix, groupRow, cell);
-    }));
-
-    it('keyboard navigation - should correct work when press tab and sft+tab on a grouped row', (async () => {
-        const fix = TestBed.createComponent(DefaultGridComponent);
-        const grid = fix.componentInstance.instance;
-
-        fix.componentInstance.width = '600px';
-        fix.componentInstance.height = '600px';
-        grid.columnWidth = '100px';
-        await wait(50);
-        fix.detectChanges();
-
-        grid.groupBy({ fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false,
-            strategy: DefaultSortingStrategy.instance() });
-        fix.detectChanges();
-
-        let groupRow = grid.groupsRowList.toArray()[0];
-        groupRow.nativeElement.dispatchEvent(new Event('focus'));
-        await wait(50);
-        fix.detectChanges();
-
-        groupRow = grid.groupsRowList.toArray()[0];
-        expect(groupRow.focused).toBe(true);
-        expect(groupRow.nativeElement.classList.contains('igx-grid__group-row--active')).toBe(true);
-        UIInteractions.triggerKeyDownEvtUponElem('tab', groupRow.nativeElement, true);
-        await wait(100);
-        fix.detectChanges();
-
-        let cell = grid.getCellByColumn(1, 'Downloads');
-        expect(cell.selected).toBe(true);
-        expect(cell.focused).toBe(true);
-
-        groupRow = grid.groupsRowList.toArray()[1];
-        groupRow.nativeElement.dispatchEvent(new Event('focus'));
-        await wait(100);
-        fix.detectChanges();
-
-        groupRow = grid.groupsRowList.toArray()[1];
-        expect(groupRow.focused).toBe(true);
-        expect(groupRow.nativeElement.classList.contains('igx-grid__group-row--active')).toBe(true);
-        groupRow.nativeElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'tab', shiftKey: true }));
-        await wait(100);
-        fix.detectChanges();
-
-        cell = grid.getCellByColumn(2, 'Released');
-        expect(cell.selected).toBe(true);
-        expect(cell.focused).toBe(true);
-    }));
-
-    it('keyboard navigation - should correct work when press tab and sft+tab on a grouped row when have row selectors', (async () => {
-        const fix = TestBed.createComponent(DefaultGridComponent);
-        const grid = fix.componentInstance.instance;
-
-        fix.componentInstance.width = '600px';
-        fix.componentInstance.height = '600px';
-        grid.columnWidth = '100px';
-        grid.rowSelectable = true;
-        await wait(30);
-        fix.detectChanges();
-
-        grid.groupBy({ fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false,
-            strategy: DefaultSortingStrategy.instance() });
-        await wait(30);
-        fix.detectChanges();
-
-        const groupRow = grid.groupsRowList.toArray()[0];
-        const firstRow = grid.getRowByIndex(1);
-        const firstRowCheckbox: HTMLElement = firstRow.nativeElement.querySelector('.igx-checkbox');
-        const cell = grid.getCellByColumn(1, 'Downloads');
-
-        groupRow.nativeElement.dispatchEvent(new Event('focus'));
-        await wait(30);
-        fix.detectChanges();
-
-        expect(groupRow.focused).toBe(true);
-        expect(groupRow.nativeElement.classList.contains('igx-grid__group-row--active')).toBe(true);
-
-        UIInteractions.triggerKeyDownEvtUponElem('tab', groupRow.nativeElement, true);
-        await wait(100);
-        fix.detectChanges();
-
-        expect(cell.selected).toBeTruthy();
-        expect(cell.focused).toBeTruthy();
-        expect(firstRowCheckbox.classList.contains('igx-checkbox--focused')).toBeFalsy();
-
-        cell.nativeElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'tab', shiftKey: true }));
-        await wait(30);
-        fix.detectChanges();
-
-        expect(cell.selected).toBeTruthy();
-        expect(cell.focused).toBeFalsy();
-        expect(firstRowCheckbox.classList.contains('igx-checkbox--focused')).toBeFalsy();
-
-        await expandCollapceGroupRow(fix, groupRow, cell);
-    }));
-
-    it('keyboard navigation - expand/colapse row with arrow keys', (async () => {
-        const fix = TestBed.createComponent(DefaultGridComponent);
-        const grid = fix.componentInstance.instance;
-
-        fix.componentInstance.width = '600px';
-        fix.componentInstance.height = '600px';
-        grid.columnWidth = '100px';
-        fix.detectChanges();
-
-        grid.groupBy({ fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false,
-            strategy: DefaultSortingStrategy.instance() });
-        fix.detectChanges();
-
-        const groupRow = grid.groupsRowList.toArray()[0];
-        groupRow.nativeElement.dispatchEvent(new Event('focus'));
-        await wait(50);
-        fix.detectChanges();
-
-        await expandCollapceGroupRow(fix, groupRow, null);
-    }));
-
-    it('keyboard navigation - should focus grouped row when press arrow keys up or down', (async () => {
-        const fix = TestBed.createComponent(DefaultGridComponent);
-        const grid = fix.componentInstance.instance;
-
-        fix.componentInstance.width = '600px';
-        fix.componentInstance.height = '600px';
-        grid.columnWidth = '100px';
-        await wait(50);
-        fix.detectChanges();
-
-        grid.groupBy({ fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false,
-            strategy: DefaultSortingStrategy.instance() });
-        fix.detectChanges();
-        let cell = grid.getCellByColumn(1, 'ID');
-        cell.nativeElement.dispatchEvent(new Event('focus'));
-        await wait(50);
-        fix.detectChanges();
-
-        expect(cell.selected).toBe(true);
-        expect(cell.focused).toBe(true);
-        UIInteractions.triggerKeyDownEvtUponElem('ArrowUp', cell.nativeElement, true);
-        await wait(100);
-        fix.detectChanges();
-
-        let groupRow = grid.groupsRowList.toArray()[0];
-        cell = grid.getCellByColumn(1, 'ID');
-        await expandCollapceGroupRow(fix, groupRow, cell);
-
-        cell = grid.getCellByColumn(2, 'ProductName');
-        cell.nativeElement.dispatchEvent(new Event('focus'));
-        await wait(100);
-        fix.detectChanges();
-
-        expect(cell.focused).toBe(true);
-        expect(cell.selected).toBe(true);
-        UIInteractions.triggerKeyDownEvtUponElem('ArrowDown', cell.nativeElement, true);
-        await wait(100);
-        fix.detectChanges();
-
-        cell = grid.getCellByColumn(2, 'ProductName');
-        groupRow = grid.groupsRowList.toArray()[1];
-
-        await expandCollapceGroupRow(fix, groupRow, cell);
-    }));
-
-    it('keyboard navigation - should correct work when press tab and sft+tab when there is a horizontal scroll', (async () => {
-        const fix = TestBed.createComponent(DefaultGridComponent);
-        const grid = fix.componentInstance.instance;
-
-        fix.componentInstance.width = '600px';
-        fix.componentInstance.height = '600px';
-        grid.columnWidth = '200px';
-        await wait(30);
-        fix.detectChanges();
-
-        grid.groupBy({ fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false,
-            strategy: DefaultSortingStrategy.instance() });
-        await wait(30);
-        fix.detectChanges();
-
-        const groupRow = grid.groupsRowList.toArray()[1];
-        const secondRow = grid.getRowByIndex(2);
-        let cell;
-
-        groupRow.nativeElement.dispatchEvent(new Event('focus'));
-        await wait(30);
-        fix.detectChanges();
-
-        expect(groupRow.focused).toBe(true);
-        groupRow.nativeElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'tab', shiftKey: true }));
-        await wait(100);
-        fix.detectChanges();
-
-        cell = grid.getCellByColumn(2, 'Released');
-        expect(cell.focused).toBe(true);
-        expect(cell.selected).toBe(true);
-
-        UIInteractions.triggerKeyDownEvtUponElem('Tab', cell.nativeElement, true);
-        await wait(100);
-        fix.detectChanges();
-        expect(cell.selected).toBe(true);
-
-        await expandCollapceGroupRow(fix, groupRow, cell);
-
-        UIInteractions.triggerKeyDownEvtUponElem('Tab', groupRow.nativeElement, true);
-        await wait(100);
-        fix.detectChanges();
-
-        cell = grid.getCellByColumn(4, 'Downloads');
-        expect(cell.focused).toBe(true);
-        expect(cell.selected).toBe(true);
-
-        cell.nativeElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'tab', shiftKey: true }));
-        await wait(30);
-        fix.detectChanges();
-
-        expect(cell.selected).toBe(true);
-        expect(groupRow.focused).toBe(true);
-    }));
-
-
-    xit('should persist last selected cell column index when navigation up through group rows.', async () => {
-        const fix = TestBed.createComponent(DefaultGridComponent);
-        const grid = fix.componentInstance.instance;
-
-        fix.componentInstance.width = '400px';
-        fix.componentInstance.height = '300px';
-        grid.columnWidth = '200px';
-        await wait();
-        fix.detectChanges();
-
-        grid.groupBy({ fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false,
-            strategy: DefaultSortingStrategy.instance() });
-        grid.groupBy({ fieldName: 'Released', dir: SortingDirection.Desc, ignoreCase: false, strategy: DefaultSortingStrategy.instance() });
-        fix.detectChanges();
-        grid.parentVirtDir.getHorizontalScroll().scrollLeft = 1000;
-        await wait(100);
-        fix.detectChanges();
-        grid.verticalScrollContainer.addScrollTop(1000);
-        await wait(200);
-        fix.detectChanges();
-        const cell = grid.getCellByColumn(20, 'Released');
-        cell.onFocus(new Event('focus'));
-        await wait(50);
-        fix.detectChanges();
-        // await HelperUtils.navigateVerticallyToIndex(grid, 20, 0, 4);
-        const row = grid.getRowByIndex(0);
-        expect(row instanceof IgxGridGroupByRowComponent).toBe(true);
-        expect(row.focused).toBe(true);
-    });
-
-    xit('should NOT clear selection from data cells when a group row is focused via KB navigation.', async () => {
-        const fix = TestBed.createComponent(DefaultGridComponent);
-        const grid = fix.componentInstance.instance;
-
-        fix.componentInstance.width = '800px';
-        fix.componentInstance.height = '300px';
-        grid.columnWidth = '200px';
-        fix.detectChanges();
-
-        grid.groupBy({ fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false,
-            strategy: DefaultSortingStrategy.instance() });
-        grid.groupBy({ fieldName: 'Released', dir: SortingDirection.Desc, ignoreCase: false, strategy: DefaultSortingStrategy.instance() });
-        fix.detectChanges();
-        const cell = grid.getCellByColumn(2, 'Downloads');
-        cell.onClick(null);
-        await wait();
-        expect(cell.selected).toBe(true);
-        await HelperUtils.navigateVerticallyToIndex(grid, 2, 0);
-
-        fix.detectChanges();
-        const row = grid.getRowByIndex(0);
-        expect(row instanceof IgxGridGroupByRowComponent).toBe(true);
-        expect(row.focused).toBe(true);
-        expect(cell.selected).toBe(true);
-    });
 
     // GroupBy + Virtualization integration
     it('should virtualize data and group records.', fakeAsync(() => {
@@ -1228,7 +716,7 @@ describe('IgxGrid - GroupBy', () => {
         fix.detectChanges();
 
         grid.groupBy({ fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false,
-            strategy: DefaultSortingStrategy.instance() });
+        strategy: DefaultSortingStrategy.instance() });
         grid.groupBy({ fieldName: 'Released', dir: SortingDirection.Desc, ignoreCase: false, strategy: DefaultSortingStrategy.instance() });
         fix.detectChanges();
 
@@ -1294,7 +782,7 @@ describe('IgxGrid - GroupBy', () => {
         fix.detectChanges();
 
         grid.groupBy({ fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false,
-            strategy: DefaultSortingStrategy.instance() });
+        strategy: DefaultSortingStrategy.instance() });
         grid.groupBy({ fieldName: 'Released', dir: SortingDirection.Desc, ignoreCase: false, strategy: DefaultSortingStrategy.instance() });
         fix.detectChanges();
 
@@ -1463,8 +951,8 @@ describe('IgxGrid - GroupBy', () => {
 
     }));
 
-    // GroupBy + Resizing
-    it('should retain same size for group row after a column is resized.', fakeAsync(() => {
+       // GroupBy + Resizing
+       it('should retain same size for group row after a column is resized.', fakeAsync(() => {
         const fix = TestBed.createComponent(DefaultGridComponent);
         const grid = fix.componentInstance.instance;
         fix.componentInstance.width = '1200px';
