+++ projects/igniteui-angular/src/lib/grids/grid/grid.pinning.spec.ts
@@ -365,133 +365,6 @@ describe('IgxGrid - Column Pinning ', () => {
         expect(headers[3].parent.parent.name).toEqual('igx-display-container');
     }));
 
-    it('should allow horizontal keyboard navigation between start pinned area and unpinned area.', fakeAsync (() => {
-        const fix = TestBed.createComponent(GridPinningComponent);
-        fix.detectChanges();
-        const grid = fix.componentInstance.instance;
-
-        fix.detectChanges();
-        tick();
-
-        grid.getColumnByName('CompanyName').pinned = true;
-        grid.getColumnByName('ContactName').pinned = true;
-
-        const cells = fix.debugElement.queryAll(By.css(CELL_CSS_CLASS));
-        let cell = cells[0];
-
-        cell.triggerEventHandler('focus', {});
-        tick();
-        fix.detectChanges();
-
-        expect(fix.componentInstance.selectedCell.value).toEqual('Maria Anders');
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('ContactName');
-
-        UIInteractions.triggerKeyDownEvtUponElem('arrowright', cell.nativeElement, true);
-        tick();
-        fix.detectChanges();
-        expect(fix.componentInstance.selectedCell.value).toEqual('Alfreds Futterkiste');
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('CompanyName');
-        cell = cells[1];
-
-        UIInteractions.triggerKeyDownEvtUponElem('arrowright', cell.nativeElement, true);
-        tick();
-
-        fix.detectChanges();
-        expect(fix.componentInstance.selectedCell.value).toEqual('ALFKI');
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('ID');
-        cell = cells[2];
-
-        UIInteractions.triggerKeyDownEvtUponElem('arrowleft', cell.nativeElement, true);
-        tick();
-        fix.detectChanges();
-        expect(fix.componentInstance.selectedCell.value).toEqual('Alfreds Futterkiste');
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('CompanyName');
-        cell.triggerEventHandler('blur', {});
-        tick();
-        cell = cells[0];
-
-        UIInteractions.triggerKeyDownEvtUponElem('arrowright', cell.nativeElement, true);
-        tick();
-        fix.detectChanges();
-        cell = cells[1];
-
-        UIInteractions.triggerKeyDownEvtUponElem('arrowright', cell.nativeElement, true);
-        tick();
-        fix.detectChanges();
-        expect(fix.componentInstance.selectedCell.value).toEqual('ALFKI');
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('ID');
-    }));
-
-    it('should allow vertical keyboard navigation in pinned area.', fakeAsync (() => {
-        const fix = TestBed.createComponent(DefaultGridComponent);
-        fix.detectChanges();
-        const grid = fix.componentInstance.instance;
-        fix.detectChanges();
-        const cells = fix.debugElement.queryAll(By.css(CELL_CSS_CLASS));
-        let cell = cells[0];
-
-        cell.triggerEventHandler('focus', {});
-
-        tick();
-        fix.detectChanges();
-
-        expect(fix.componentInstance.selectedCell.value).toEqual('Alfreds Futterkiste');
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('CompanyName');
-
-        UIInteractions.triggerKeyDownEvtUponElem('arrowdown', cell.nativeElement, true);
-
-        tick();
-        grid.cdr.detectChanges();
-
-        expect(fix.componentInstance.selectedCell.value).toEqual('Ana Trujillo Emparedados y helados');
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('CompanyName');
-        cell = cells[6];
-
-        UIInteractions.triggerKeyDownEvtUponElem('arrowup', cell.nativeElement, true);
-
-        tick();
-        grid.cdr.detectChanges();
-
-        expect(fix.componentInstance.selectedCell.value).toEqual('Alfreds Futterkiste');
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('CompanyName');
-    }));
-
-    it('should allow keyboard navigation to first/last cell with Ctrl when there are the pinned columns.', async () => {
-        const fix = TestBed.createComponent(GridPinningComponent);
-        fix.detectChanges();
-
-        await wait();
-        const grid = fix.componentInstance.instance;
-        grid.getColumnByName('CompanyName').pinned = true;
-        grid.getColumnByName('ContactName').pinned = true;
-        fix.detectChanges();
-        const cells = fix.debugElement.queryAll(By.css(CELL_CSS_CLASS));
-        let cell = cells[0];
-
-        cell.triggerEventHandler('focus', {});
-        await wait();
-        fix.detectChanges();
-
-        expect(fix.componentInstance.selectedCell.value).toEqual('Maria Anders');
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('ContactName');
-
-        cell.nativeElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'arrowright', ctrlKey: true }));
-        await wait(30);
-        fix.detectChanges();
-
-        expect(fix.componentInstance.selectedCell.value).toEqual('030-0076545');
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('Fax');
-
-        cell = cells[cells.length - 1];
-        cell.nativeElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'arrowleft', ctrlKey: true }));
-        await wait(30);
-        fix.detectChanges();
-
-        // It won't scroll left since the next selected cell will be in the pinned area
-        expect(fix.componentInstance.selectedCell.value).toEqual('Maria Anders');
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('ContactName');
-    });
-
     it('should allow hiding/showing pinned column.', fakeAsync(() => {
         const fix = TestBed.createComponent(GridPinningComponent);
         fix.detectChanges();
